Change the expression of recording system sound option
Previously, the app only records system sound when you use this option.
However, after #42 now it records both system sound and mic sound. I
changed the label of this option from "Record system sound" to "Include
system sound" so as not to cause a misunderstanding that using this
option only records system sound